refactor(search): moved helper funcs to utils.go
